Added setting for a "dimming type."
Added uint8 to FlashData to store the dimming type (and 24 extra bytes
so the entire structure is a multiple of 32bits - else it won't save.)
The settings page button just cycles through the currently support
"dimming types" which are to always dim (default), never dim, and only
dim when the printer status is idle.  Non-English translations were via
google translate.