Make build process more host architecture independent
This is mnostly for the 5i/7i builds that tried to use the Windows
"copy" command to append the splash screen to the binary...

Added a makefile.targets file that will build the bin and -nologo.bin
files from the elf and splash screen.  If the OS is Windows_NT,
the "copy" command will be used, if not, "cat" will be used.

Updated the build configuration for the 5i/7i in the project settings
to do a make on the bin and -nologo.bin files instead of calling "copy"
directly.

Also updated the Build Instructions to add a section for Linux.