Add "off" to the tool button press cycle
The bed and tool selection buttons now cycle
off -> active-> standby ->off.  The Active and
Standby temp buttons will indicate which state
the tool/heater is in.

Since turning a heater off requires setting its
temp to -273, we keep the last real active and
standby temps locally in the activeTemps
and standbyTemps arrays so the user doesn't
have to re-enter them.