Fix initialization of Macro Columns setting and revery sticky temps
The Macro Columns setting was being stored as an unsigned char
when it should have been signed.  This caused the panel to try
and display 255 columns.

Also reverted the "sticky temp" feature for now.  This feature
saved the last temp a heater was set at when it was turned off
to allow easy restoral. Unfortunately, that's different than
what the DWC does.  Will consider re-adding it with a setup
option to control it at a later date.