Added sticky temps back and fixed issues with macro columns
The ability to remember the active and standby temps when
heaters are turned off is back and controlled by a new
setup option.  The default is "off" to be compatible with
the DWC.

Fixed an issue where the tool selection button wasn't
cycling through the states correctly.

Fixed an issue where the number of macro columns
wasn't taking effect until a panel reset