Block evSelectHead while printing
If a button for selecting/deselecting a tool (heater/head) is pressed mid-print, it's almost always an accidental press, but always fatal to the print.  So, don't react to the button while printing.  If the user knows what they are doing, they can use the console to manually send "T" gcode commands.